+++ lib/jekyll-seo-tag.rb
@@ -30,7 +30,7 @@ module Jekyll
     def options
       {
         "version" => Jekyll::SeoTag::VERSION,
-        "title"   => title?
+        "title"   => title?,
       }
     end
 
@@ -39,18 +39,18 @@ module Jekyll
         "page"      => context.registers[:page],
         "site"      => context.registers[:site].site_payload["site"],
         "paginator" => context["paginator"],
-        "seo_tag"   => options
+        "seo_tag"   => options,
       }
     end
 
     def title?
-      !(@text =~ %r!title=false!i)
+      @text !~ %r!title=false!i
     end
 
     def info
       {
         :registers => context.registers,
-        :filters   => [Jekyll::Filters]
+        :filters   => [Jekyll::Filters],
       }
     end